requirements.txt
requests==2.32.3
fastapi==0.115.5
uvicorn==0.32.0
python-dotenv==1.0.1
loguru==0.7.2
